refactor storable map files
